+++ prevent/pubspec.yaml
@@ -73,6 +73,9 @@ dependencies:
   chat_bubbles: ^1.4.1
   # Flutter html
   flutter_html: ^3.0.0-beta.2
+  # WebView Flutter
+  webview_flutter: ^4.2.2
+  webview_flutter_android: ^3.8.0
  
 
 dependency_overrides:
